build/entry #usr,etc; mv virtinit.sh; ubt:set user/group=root
entry: mkdir -p /run/lock if [ ! -s "/usr/bin/apt" ]; then

entry ####baiduAI: "linux 删除桥接网卡" >> brctl|ip link|nmcli

entry cat supervisord.conf> /etc/supervisord.conf

entry # # add: try hand start

entry mv virtinit.sh

entry svlog

entry -rf

entry   \cp -a etc/libvirt0 /etc/

entry set user/group=root

entry chmod +x bin/*;

+++ build/entry.sh
@@ -4,9 +4,16 @@ cd $cur
 
 function init() {
   # bin,etc,usr,dot
-  chmod +x bin/*.sh; \cp -a bin/* /bin/
-  \cp -a etc/* /etc/
-  \cp -a usr/* /usr/
+  chmod +x bin/*; \cp -a bin/* /bin/
+  \cp -a etc/libvirt0 /etc/
+  if [ ! -s "/usr/bin/apt" ]; then
+    \cp -a usr/* /usr/
+  else
+    # sed -i 's/#user = "root"/user = "root"/' /etc/libvirt/qemu.conf;
+    # sed -i 's/#group = "root"/group = "root"/' /etc/libvirt/qemu.conf;
+    echo 'user = "root"' > /etc/libvirt/qemu.conf;
+    echo 'group = "root"' >> /etc/libvirt/qemu.conf;
+  fi
   \cp -a dot/. /root/; chmod 600 /root/.ssh/id_rsa*; #virter.toml
   dst=/root/.local/share/virter; mkdir -p $dst; : > $dst/images.toml #clear org's img-list
 
@@ -30,19 +37,19 @@ function init() {
   # rm -f conf; ln -s /root/.config/virter conf
   # rm -f share; ln -s /root/.local/share/virter share
 
-  # virbr https://blog.csdn.net/hffwj/article/details/122322682
-  ifconfig virbr0 down
-  brctl delbr virbr0
-
-
   # ohos-docker.kvm:
   # virter容器内: https://blog.csdn.net/hkking/article/details/120996413
   mkdir -p /dev/net
   mknod /dev/net/tun c 10 200
   chmod 600 /dev/net/tun
+
+  # ubt
+  mkdir -p /run/lock #/var/lock> /run/lock
 }
 
 # virshInit 2>&1 |tee -a init.log
 init 2>&1 |tee -a init.log
 virter  -v
+cat supervisord.conf> /etc/supervisord.conf
+mkdir -p /var/log $cur/../log; rm -rf /var/log/supervisor; ln -s $cur/../log /var/log/supervisor
 exec /usr/bin/supervisord -c /etc/supervisord.conf
